casper (lib) :: tItem tabs type updated to usize

+++ inter_lang/src/ast/s_key.rs
@@ -15,7 +15,7 @@ impl sKey {
 }
 
 impl tItem for sKey {
-    fn as_string(&self, tabs: u8, indent_first_line: bool) -> String {
+    fn as_string(&self, tabs: usize, indent_first_line: bool) -> String {
         let mut as_string = "".to_string();
 
         if indent_first_line {
